+++ Libplanet.Stun/Stun/TurnClient.cs
@@ -7,6 +7,7 @@ using System.Net;
 using System.Net.Sockets;
 using System.Threading;
 using System.Threading.Tasks;
+using Libplanet.Net;
 using Libplanet.Stun.Messages;
 using Nito.AsyncEx;
 using Serilog;
@@ -215,6 +216,31 @@ namespace Libplanet.Stun
             }
             catch (IOException)
             {
+            }
+        }
+
+        public async Task BindProxies(
+            int listenPort,
+            CancellationToken cancellationToken = default(CancellationToken))
+        {
+            while (!cancellationToken.IsCancellationRequested)
+            {
+#pragma warning disable IDE0067 // We'll dispose of `stream` in proxy task.
+                NetworkStream stream = await AcceptRelayedStreamAsync(cancellationToken);
+#pragma warning restore IDE0067
+
+                // TODO We should expose the interface so that library users
+                // can limit / manage the task.
+                Func<Task> startAsync = async () =>
+                {
+                    using var proxy = new NetworkStreamProxy(stream);
+                    await proxy.StartAsync(IPAddress.Loopback, listenPort);
+                };
+
+#pragma warning disable CS4014
+                Task.Run(startAsync, cancellationToken)
+                    .ContinueWith(_ => stream.Dispose(), cancellationToken);
+#pragma warning restore CS4014
             }
         }